Split to 'sound-notification' store & get the correct full_name & subject title
