+++ src/js/actions/actions.js
@@ -5,6 +5,7 @@ var Actions = Reflux.createActions({
   'login': {},
   'logout': {},
   'getNotifications': {asyncResult: true},
+  'isNewNotification': {},
   'updateSearchTerm': {},
   'clearSearchTerm': {},
   'setSetting': {}
